Moved matrix transform logic to geom.js
