+++ src/path.js
@@ -1,3 +1,5 @@
+import { matrixTransform, transformEllipse } from './geom'
+
 let Path = (init) => {
   let instructions = init || []
 
@@ -53,71 +55,6 @@ let Path = (init) => {
         return [params[0], params[1]]
       case 'A':
         return [params[5], params[6]]
-    }
-  }
-
-  let matrixTransform = (points, m) => {
-    return points.map(point => {
-      return {
-        x: point.x * m[0] + point.y * m[2] + m[4],
-        y: point.x * m[1] + point.y * m[3] + m[5]
-      }
-    })
-  }
-
-  let transformEllipse = (rx, ry, ax, m) => {
-    const torad = Math.PI / 180
-    const epsilon = 0.0000000001
-
-    let c = Math.cos(ax * torad), s = Math.sin(ax * torad)
-    let ma = [
-      rx * (m[0]*c + m[2]*s),
-      rx * (m[1]*c + m[3]*s),
-      ry * (-m[0]*s + m[2]*c),
-      ry * (-m[1]*s + m[3]*c)
-    ]
-
-    let J = ma[0]*ma[0] + ma[2]*ma[2],
-      K = ma[1]*ma[1] + ma[3]*ma[3]
-    
-    let D = ((ma[0]-ma[3])*(ma[0]-ma[3]) + (ma[2]+ma[1])*(ma[2]+ma[1])) *
-      ((ma[0]+ma[3])*(ma[0]+ma[3]) + (ma[2]-ma[1])*(ma[2]-ma[1]))
-
-    let JK = (J + K) / 2
-    
-    if (D < epsilon * JK) {
-      return {
-        rx: Math.sqrt(JK),
-        ry: Math.sqrt(JK),
-        ax: 0,
-        isDegenerate: false
-      }
-    }
-
-    let L = ma[0]*ma[1] + ma[2]*ma[3]
-    D = Math.sqrt(D)
-
-    let l1 = JK + D/2,
-      l2 = JK - D/2
-
-    let newAx, newRx, newRy
-    newAx = (Math.abs(L) < epsilon && Math.abs(l1 - K) < epsilon) ? 90
-      : Math.atan(Math.abs(L) > Math.abs(l1 - K) ? (l1 - J) / L : L / (l1 - K)) * 180 / Math.PI
-
-    if (newAx >= 0) {
-      newRx = Math.sqrt(l1)
-      newRy = Math.sqrt(l2)
-    } else {
-      newAx += 90;
-      newRx = Math.sqrt(l2)
-      newRy = Math.sqrt(l1)
-    }
-
-    return {
-      rx: newRx,
-      ry: newRy,
-      ax: newAx,
-      isDegenerate: (newRx < epsilon * newRy || newRy < epsilon * newRx)
     }
   }
